Change augur-app to use augur-node exported controller. Dropped about 100 lines, I think we could drop some more. The main benefit here is that a lot of the behavior from augur-app is now handled inside augur-node and augur-app doesn't reach so far into the internals of augur-node
